testRemoveGhost fault
