--- src/main/java/dev/lidia/karina/ghost_buster/App.java
@@ -1,18 +1,0 @@
-package dev.lidia.karina.ghost_buster;
-
-import dev.lidia.karina.ghost_buster.model.Container;
-
-public final class App {
-    private App() {
-    }
-
-    
-    
-    public class Main {
-        public  void main(String[] args) {
-           
-        }
-    }
-       
-    
-}
