+++ src/test/java/dev/lidia/karina/ghost_buster/model/ContainerTest.java
@@ -29,16 +29,17 @@ public class ContainerTest {
 
     @Test
     public void testGhostList() {
-        ghosts.forEach(ghost -> {
-            System.out.println("ID: " + ghost.getId());
-            System.out.println("Nombre: " + ghost.getName());
-            System.out.println("Tipo: " + ghost.getType());
-            System.out.println("Nivel: " + ghost.getLevel());
-            System.out.println("Fecha: " + ghost.getDate());
-            System.out.println("Habilidad: " + ghost.getAbility());
-            System.out.println("Afinidad: " + ghost.getAfinity());
-            System.out.println("----------------------------------------");
-        });
+        Container container = new Container();
+        container.addGhost(new SimulatedGhost(1, "Fantasma1", "TipoA", "Nivel1", "2025-01-01", "Habilidad1", "Afinidad1"));
+        container.addGhost(new SimulatedGhost(2, "Fantasma2", "TipoB", "Nivel2", "2025-01-02", "Habilidad2", "Afinidad2"));
+        container.ghostList(); 
+    }
+
+    @Test
+    public void testRemoveGhost() {
+        container.addGhost(ghost1);
+        container.removeGhost(1);
+        assertEquals(0, container.getGhosts().size());
     }
 
 }
